requirements.txt
fastapi
uvicorn
pandas
requests
google-cloud-storage
pydantic
